+++ CMakeLists.txt
@@ -1634,6 +1634,7 @@ if(USE_SE_V8)
                      cocos/bindings/jswrapper/v8/SeApi.h
         NO_WERROR    cocos/bindings/jswrapper/v8/Utils.cpp
                      cocos/bindings/jswrapper/v8/Utils.h
+                     cocos/bindings/jswrapper/v8/MissingSymbols.h
     )
     if(USE_V8_DEBUGGER)
     cocos_source_files(MODULE ccbindings
